requirements.txt
numpy
pandas
tensorflow
matplotlib
scikit-learn
scipy
strupnet
pyserial